requirements.txt
boto3==1.26.147
Django==4.2.2
django-environ==0.10.0
djangorestframework==3.12.2
psycopg2==2.9.6